restyled a bunch, fix teacher navigations to the right screens, fixed lightstatusbar component
