+++ App.tsx
@@ -2,6 +2,7 @@ import React from "react";
 import MyStack from "./src/navigation/stack";
 import { useFonts } from "expo-font";
 import StatusBarLight from "./src/components/StatusBarLight";
+import StudentList from "./src/screens/TeacherRole/StudentList";
 
 export default function App() {
   const [fontsLoaded] = useFonts({
