+++ app.json
@@ -8,7 +8,7 @@
     "userInterfaceStyle": "light",
     "splash": {
       "image": "./assets/splash1.png",
-      "resizeMode": "fill",
+      "resizeMode": "cover",
       "backgroundColor": "#FFFFFF"
     },
     "updates": {
